requirements.txt
aniso8601==9.0.1
blinker==1.7.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-RESTful==0.3.10
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
munch==4.0.0
mutagen==1.47.0
psycopg==3.1.15
psycopg-binary==3.1.15
pycryptodomex==3.19.0
Pygments==2.17.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
shellingham==1.5.4
six==1.16.0
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
websockets==12.0
Werkzeug==3.0.1
yt-dlp==2023.11.16
